Move log and config into common
