+++ internal/core/src/CMakeLists.txt
@@ -20,11 +20,9 @@ include_directories(${MILVUS_THIRDPARTY_SRC})
 set(FOUND_OPENBLAS "unknown")
 add_subdirectory( exceptions )
 add_subdirectory( utils )
-add_subdirectory( log )
 add_subdirectory( pb )
 add_subdirectory( segcore )
 add_subdirectory( query )
 add_subdirectory( common )
 add_subdirectory( indexbuilder )
-add_subdirectory( config )
 add_subdirectory( index )
