+++ internal/core/src/common/Log.h
