+++ internal/core/src/segcore/reduce_c.cpp
@@ -16,10 +16,10 @@
 #include "Reduce.h"
 #include "common/CGoHelper.h"
 #include "common/Consts.h"
+#include "common/Log.h"
 #include "common/Types.h"
 #include "common/QueryResult.h"
 #include "exceptions/EasyAssert.h"
-#include "log/Log.h"
 #include "pb/milvus.pb.h"
 #include "query/Plan.h"
 #include "segcore/ReduceStructure.h"
