+++ internal/core/src/segcore/segcore_init_c.cpp
@@ -9,8 +9,8 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
-#include "config/ConfigKnowhere.h"
-#include "log/Log.h"
+#include "common/ConfigKnowhere.h"
+#include "common/Log.h"
 #include "segcore/SegcoreConfig.h"
 #include "segcore/segcore_init_c.h"
