+++ internal/core/unittest/CMakeLists.txt
@@ -72,7 +72,6 @@ if (LINUX)
             milvus_segcore
             milvus_indexbuilder
             milvus_index
-            milvus_log
             pthread
             )
     install(TARGETS index_builder_test DESTINATION unittest)
@@ -87,7 +86,6 @@ target_link_libraries(all_tests
         milvus_segcore
         milvus_indexbuilder
         milvus_index
-        milvus_log
         pthread
         )
